{{cookiecutter.project_slug}}/requirements.txt
# external requirements
pip==19.2.3
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
flake8==3.7.8
tox==3.14.0
coverage==4.5.4
python-dotenv
Sphinx==1.8.5
twine==1.14.0
awscli
{% if cookiecutter.command_line_interface|lower == 'click' -%}
Click==7.0{% endif %}
{% if cookiecutter.use_pytest == 'y' -%}
pytest==4.6.5
pytest-runner==5.1{% endif %}
{% if cookiecutter.python_interpreter != 'python3' %}
# backwards compatibility
pathlib2
{% endif %}